Push for Final Submission. See notes.
Inputs are actually not rebindable at the moment due to some issue with the new Input system so that is not a feature. Few known bugs including leftover interaction prompt when going to level, some water tiles in level are wrong or missing on rare occasion. Multitude of changes and new features aswell, stackable items, healing, death and AI, looting, just to list a few.